==> _episodes_rmd/06-data-structures-part2.Rmd <==
---
title: "Exploring Data Frames"
teaching: 20
exercises: 10
questions:
- "How can I manipulate a data frame?"
objectives:
- "Add and remove rows or columns."
- "Remove rows with `NA` values."
- "Append two data frames."
- "Understand what a `factor` is."
- "Convert a `factor` to a `character` vector and vice versa."
- "Display basic properties of data frames including size and class of the columns, names, and first few rows."
keypoints:
- "Use `cbind()` to add a new column to a data frame."
- "Use `rbind()` to add a new row to a data frame."
- "Remove rows from a data frame."
- "Use `na.omit()` to remove rows from a data frame with `NA` values."
- "Use `levels()` and `as.character()` to explore and manipulate factors."
- "Use `str()`, `summary()`, `nrow()`, `ncol()`, `dim()`, `colnames()`, `rownames()`, `head()`, and `typeof()` to understand the structure of a data frame."
- "Read in an Excel file using read_excel()`."
- "Understand what `length()` of a data frame represents."
source: Rmd
---

```{r, include=FALSE}
source("../bin/chunk-options.R")
knitr_fig_path("06-")
```

At this point, you've seen it all: in the last lesson, we toured all the basic
data types and data structures in R. Everything you do will be a manipulation of
those tools. But most of the time, the star of the show is the data frame—the table that we created by loading information from a csv file. In this lesson, we'll learn a few more things
about working with data frames.

## Adding columns and rows in data frames

We already learned that the columns of a data frame are vectors, so that our
data are consistent in type throughout the columns. As such, if we want to add a
new column, we can start by making a new vector:

```{r, echo = FALSE}
cats <- read.csv("data/feline-data.csv", stringsAsFactors = TRUE)
```

```{r}
age <- c(2, 3, 5)
cats
```

We can then add this as a column via:

```{r}
cbind(cats, age)
```

Note that if we tried to add a vector of ages with a different number of entries than the number of rows in the data frame, it would fail:

```{r, error=TRUE}
age <- c(2, 3, 5, 12)
cbind(cats, age)

age <- c(2, 3)
cbind(cats, age)
```

Why didn't this work? Of course, R wants to see one element in our new column
for every row in the table:

```{r}
nrow(cats)
length(age)
```

So for it to work we need to have `nrow(cats)` = `length(age)`. Let's overwrite the content of cats with our new data frame.

```{r}
age <- c(2, 3, 5)
cats <- cbind(cats, age)
```

Now how about adding rows? We already know that the rows of a
data frame are lists:

```{r}
newRow <- list("tortoiseshell", 3.3, TRUE, 9)
cats <- rbind(cats, newRow)
```

Looks like our attempt to use the `rbind()` function returns a warning.  Recall that, unlike errors, warnings do not necessarily stop a function from performing its intended action.  You can confirm this by taking a look at the `cats` data frame.

```{r}
cats
```

Notice that not only did we successfully add a new row, but there is `NA` in the column *coats* where we expected "tortoiseshell" to be.  Why did this happen?

## Factors

For an object containing the data type `factor`, each different value represents what is called a `level`. In our case, the `factor` "coat" has 3 levels: "black", "calico", and "tabby". R will only accept values that match one of the levels. If you add a new value, it will become `NA`.

The warning is telling us that we unsuccessfully added "tortoiseshell" to our
*coat* factor, but 3.3 (a numeric), TRUE (a logical), and 9 (a numeric) were
successfully added to *weight*, *likes_string*, and *age*, respectively, since
those variables are not factors. To successfully add a cat with a
"tortoiseshell" *coat*, add "tortoiseshell" as a possible *level* of the factor:

```{r}
levels(cats$coat)
levels(cats$coat) <- c(levels(cats$coat), "tortoiseshell")
cats <- rbind(cats, list("tortoiseshell", 3.3, TRUE, 9))
```

Alternatively, we can change a factor into a character vector; we lose the
handy categories of the factor, but we can subsequently add any word we want to the
column without babysitting the factor levels:

```{r}
str(cats)
cats$coat <- as.character(cats$coat)
str(cats)
```

> ## Challenge 1
> Let's imagine that 1 cat year is equivalent to 7 human years.
> 1. Create a vector called `human_age` by multiplying `cats$age` by 7.
> 2. Convert `human_age` to a factor.
> 3. Convert `human_age` back to a numeric vector using the `as.numeric()` function. Now divide it by 7 to get the original ages back. Explain what happened.
>
> > ## Solution to Challenge 1
> > 1. `human_age <- cats$age * 7`
> > 2. `human_age <- factor(human_age)`. `as.factor(human_age)` works just as well.
> > 3. `as.numeric(human_age)` yields `1 2 3 4 4` because factors are stored as integers (here, 1:4), each of which is associated with a label (here, 28, 35, 56, and 63). Converting the factor to a numeric vector gives us the underlying integers, not the labels. If we want the original numbers, we need to convert `human_age` to a character vector (using `as.character(human_age)`) and then to a numeric vector (why does this work?). This comes up in real life when we accidentally include a character somewhere in a column of a .csv file supposed to only contain numbers, and forget to set `stringsAsFactors=FALSE` when we read in the data.
> {: .solution}
{: .challenge}

## Removing rows

We now know how to add rows and columns to our data frame in R—but in our
first attempt to add a "tortoiseshell" cat to the data frame we have accidentally
added a garbage row:

```{r}
cats
```

We can ask for a data frame minus this offending row:

```{r}
cats[-4, ]
```

Notice the comma with nothing after it to indicate that we want to drop the entire fourth row.

Note: we could also remove both new rows at once by putting the row numbers
inside of a vector: `cats[c(-4,-5), ]`

Alternatively, we can drop all rows with `NA` values:

```{r}
na.omit(cats)
```

Let's reassign the output to `cats`, so that our changes will be permanent:

```{r}
cats <- na.omit(cats)
```

## Removing columns  

We can also remove columns in our data frame. What if we want to remove the column "age". We can remove it in two ways, by variable number or by index.  

```{r}
cats[,-4]
```

Notice the comma with nothing before it, indicating we want to keep all of the rows.  

Alternatively, we can drop the column by using the index name and the `%in%` operator. The `%in%` operator goes through each element of its left argument, in this case the names of `cats`, and asks, "Does this element occur in the second argument?"

```{r}
drop <- names(cats) %in% c("age")
cats[,!drop]
```

We will cover subsetting with logical operators like `%in%` in more detail in the next episode. See the section [Subsetting through other logical operations](../06-data-subsetting/#logical-operations)

## Appending to a data frame

The key to remember when adding data to a data frame is that *columns are
vectors and rows are lists.* We can also glue two data frames
together with `rbind`:

```{r}
cats <- rbind(cats, cats)
cats
```
But now the row names are unnecessarily complicated. We can remove the rownames,
and R will automatically re-name them sequentially:

```{r}
rownames(cats) <- NULL
cats
```

> ## Challenge 2
>
> You can create a new data frame right from within R with the following syntax:
> ```{r}
> df <- data.frame(id = c("a", "b", "c"),
>                  x = 1:3,
>                  y = c(TRUE, TRUE, FALSE),
>                  stringsAsFactors = FALSE)
> ```
> Make a data frame that holds the following information for yourself:
>
> - first name
> - last name
> - lucky number
>
> Then use `rbind` to add an entry for the people sitting beside you.
> Finally, use `cbind` to add a column with each person's answer to the question, "Is it time for coffee break?"
>
> > ## Solution to Challenge 2
> > ```{r}
> > df <- data.frame(first = c("Grace"),
> >                  last = c("Hopper"),
> >                  lucky_number = c(0),
> >                  stringsAsFactors = FALSE)
> > df <- rbind(df, list("Marie", "Curie", 238) )
> > df <- cbind(df, coffeetime = c(TRUE,TRUE))
> > ```
> {: .solution}
{: .challenge}

## Realistic example
So far, you have seen the basics of manipulating data frames with our cat data;
now let's use those skills to digest a more realistic dataset. Let's read in the
`gapminder` dataset like we downloaded previously, but this time, we'll use the
[readxl](https://cran.r-project.org/package=readxl) library
to import an Excel version.

Use the searchbox in the `Packages` tab at the bottom right to check if `readxl`
is installed. If not, click `Install` then search for `readxl` and install before
loading the package.

Finally, you will need to download the XLSX data from [this link](https://github.com/ndporter/dspg-r-2022/raw/gh-pages/data/gapminder_data.xlsx).

```{r}
library(readxl)
gapminder <- read_excel("data/gapminder_data.xlsx")
```

The `read_excel` function will automatically guess the format (xls or xlsx) and choose the first worksheet. Check the help file to see how to change these.

> ## Miscellaneous Tips
>
> * Another type of file you might encounter are tab-separated value files (.tsv). To specify a tab as a separator, use `"\\t"` or `read.delim()`.
>
> * Files can also be downloaded directly from the Internet into a local
> folder of your choice onto your computer using the `download.file` function.
> The `read.csv` function can then be executed to read the downloaded file from the download location, for example,
> ```{r eval=FALSE, echo=TRUE}
> download.file("https://raw.githubusercontent.com/swcarpentry/r-novice-gapminder/gh-pages/_episodes_rmd/data/gapminder_data.csv", destfile = "data/gapminder_data.csv")
> gapminder <- read.csv("data/gapminder_data.csv", stringsAsFactors = TRUE)
> ```
>
> * Alternatively, you can also read in files directly into R from the Internet by replacing the file paths with a web address in `read.csv`. One should note that in doing this no local copy of the csv file is first saved onto your computer. For example,
> ```{r eval=FALSE, echo=TRUE}
> gapminder <- read.csv("https://raw.githubusercontent.com/swcarpentry/r-novice-gapminder/gh-pages/_episodes_rmd/data/gapminder_data.csv", stringsAsFactors = TRUE)
> ```
{: .callout}

Let's investigate gapminder a bit; the first thing we should always do is check
out what the data looks like with `str`:

```{r}
str(gapminder)
```

Unlike before, `country` and `continent` are character vectors. To treat them as factors, we need to either use the `col_types` option for `read_excel` or change them after creating the dataframe.

```{r}
gapminder$country <- factor(gapminder$country)
gapminder$continent <- factor(gapminder$continent)
```

An additional method for examining the structure of gapminder is to use the `summary` function. This function can be used on various objects in R. For data frames, `summary` yields a numeric, tabular, or descriptive summary of each column. Factor columns are summarized by the number of items in each level, numeric or integer columns by the descriptive statistics (quartiles and mean), and character columns by its length, class, and mode.

```{r}
summary(gapminder$country)
```

Along with the `str` and `summary` functions, we can examine individual columns of the data frame with our `typeof` function:

```{r}
typeof(gapminder$year)
typeof(gapminder$country)
str(gapminder$country)
```

We can also interrogate the data frame for information about its dimensions;
remembering that `str(gapminder)` said there were 1704 observations of 6
variables in gapminder, what do you think the following will produce, and why?

```{r}
length(gapminder)
```

A fair guess would have been to say that the length of a data frame would be the
number of rows it has (1704), but this is not the case; remember, a data frame
is a *list of vectors and factors*:

```{r}
typeof(gapminder)
```

When `length` gave us 6, it's because gapminder is built out of a list of 6
columns. To get the number of rows and columns in our dataset, try:

```{r}
nrow(gapminder)
ncol(gapminder)
```

Or, both at once:

```{r}
dim(gapminder)
```

We'll also likely want to know what the titles of all the columns are, so we can
ask for them later:

```{r}
colnames(gapminder)
```

At this stage, it's important to ask ourselves if the structure R is reporting
matches our intuition or expectations; do the basic data types reported for each
column make sense? If not, we need to sort any problems out now before they turn
into bad surprises down the road, using what we've learned about how R
interprets data, and the importance of *strict consistency* in how we record our
data.

Once we're happy that the data types and structures seem reasonable, it's time
to start digging into our data proper. Check out the first few lines:

```{r}
head(gapminder)
```

> ## Challenge 3
>
> It's good practice to also check the last few lines of your data and some in the middle. How would you do this?
>
> Searching for ones specifically in the middle isn't too hard, but we could ask for a few lines at random. How would you code this?
>
> > ## Solution to Challenge 3
> > To check the last few lines it's relatively simple as R already has a function for this:
> >
> > ~~~
> > tail(gapminder)
> > tail(gapminder, n = 15)
> > ~~~
> > {: .r}
> >
> > What about a few arbitrary rows just in case something is odd in the middle?
> >
> > ## Tip: Sampling Random Rows
> > The solution here presents one form of using nested functions, i.e. a function passed as an argument to another function. This might sound like a new concept, but you are already using it!
> >
> > Remember my_dataframe[rows, cols] will print to screen your data frame with the number of rows and columns you asked for (although you might have asked for a range or named columns for example). How would you get the last row if you don't know how many rows your data frame has? R has a function for this. 
> >
> > What about getting a (pseudorandom) sample? R also has a function for this.
> > ~~~
> > gapminder[sample(nrow(gapminder), 5), ]
> > ~~~
> > {: .r}
> {: .solution}
{: .challenge}


To make sure our analysis is reproducible, we should put the code
into a script file so we can come back to it later.

> ## Challenge 4
>
> Go to file -> new file -> R script, and write an R script
> to load in the gapminder dataset. Put it in the `scripts/`
> directory and add it to version control.
>
> Run the script using the `source` function, using the file path
> as its argument (or by pressing the "source" button in RStudio).
>
> > ## Solution to Challenge 4
> >
> > The `source` function can be used to use a script within a script.
> > Assume you would like to load the same type of file over and over
> > again and therefore you need to specify the arguments to fit the
> > needs of your file. Instead of writing the necessary argument again
> > and again you could just write it once and save it as a script. Then,
> > you can use `source("Your_Script_containing_the_load_function")` in a new
> > script to use the function of that script without writing everything again.
> > Check out `?source` to find out more.
> >
> > ```{r eval = FALSE}
> > library(readxl)
> > gapminder <- read_excel(file = "data/gapminder_data.xlsx")
> > gapminder$country <- factor(gapminder$country)
> > gapminder$continent <- factor(gapminder$continent)
> > ```
> > Once the script is ready, save it in `scripts/` as `load-gapminder.R`.
> > 
> > To run the script and load the data into the `gapminder` variable:
> > ```{r eval = FALSE}
> > source(file = "scripts/load-gapminder.R")
> > ```
> {: .solution}
{: .challenge}

> ## Challenge 5
>
> Read the output of `str(gapminder)` again;
> this time, use what you've learned about factors, lists and vectors,
> as well as the output of functions like `colnames` and `dim`
> to explain what everything that `str` prints out for gapminder means.
> If there are any parts you can't interpret, discuss with your neighbors!
>
> > ## Solution to Challenge 5
> >
> > The object `gapminder` is a data frame with columns
> > - `country` and `continent` are factors.
> > - `year` is an integer vector.
> > - `pop`, `lifeExp`, and `gdpPercap` are numeric vectors.
> >
> {: .solution}
{: .challenge}
