==> _episodes_rmd/12-wrap-up.Rmd <==
---
title: Writing Good Software
teaching: 15
exercises: 0
questions:
- "How can I write software that other people can use?"
objectives:
- "Describe best practices for writing R and explain the justification for each."
keypoints:
- "Keep your project folder structured, organized and tidy."
- "Document what and why, not how."
- "Break programs into short single-purpose functions."
- "Write re-runnable tests."
- "Don't repeat yourself."
- "Be consistent in naming, indentation, and other aspects of style."
source: Rmd
---

## Structure your project folder

Keep your project folder structured, organized and tidy, by creating subfolders for your code files, manuals, data, binaries, output plots, etc. It can be done completely manually, or with the help of RStudio's `New Project` functionality, or a designated package, such as [`ProjectTemplate`](http://projecttemplate.net/index.html).

## Make code readable

The most important part of writing code is making it readable and understandable.
You want someone else to be able to pick up your code and be able to understand
what it does: more often than not this someone will be you 6 months down the line,
who will otherwise be cursing past-self.

## Documentation: tell us what and why, not how

When you first start out, your comments will often describe what a command does,
since you're still learning yourself and it can help to clarify concepts and
remind you later. However, these comments aren't particularly useful later on
when you don't remember what problem your code is trying to solve. Try to also
include comments that tell you *why* you're solving a problem, and *what* problem
that is. The *how* can come after that: it's an implementation detail you ideally
shouldn't have to worry about.

## Keep your code modular

Our recommendation is that you should separate your functions from your analysis
scripts, and store them in a separate file that you `source` when you open the R
session in your project. This approach is nice because it leaves you with an
uncluttered analysis script, and a repository of useful functions that can be
loaded into any analysis script in your project. It also lets you group related
functions together easily.

## Break down problem into bite size pieces

When you first start out, problem solving and function writing can be daunting
tasks, and hard to separate from code inexperience. Try to break down your
problem into digestible chunks and worry about the implementation details later:
keep breaking down the problem into smaller and smaller functions until you
reach a point where you can code a solution, and build back up from there.

## Know that your code is doing the right thing

Make sure to test your functions!

## Don't repeat yourself

Functions enable easy reuse within a project. If you see blocks of similar
lines of code through your project, those are usually candidates for being
moved into functions.

If your calculations are performed through a series of functions, then the
project becomes more modular and easier to change. This is especially the case
for which a particular input always gives a particular output.

## Remember to be stylish

Apply consistent style to your code.

## Make regular commits with a single focus

When using version control to track changes, it's much simpler to see what happened
(and reverse it, if necessary) if you regularly add files and commit changes.
Try to make each commit focused on a single goal or purpose - if you can't describe
what you did in around 100 characters (the commit message), consider splitting it
into multiple commits. This also helps discover dependency issues, where one change
impacts other code, soon enough to find and resolve them quickly without searching
dozens of updated files.

You can commit changes locally and even push them to your branch of a project, even
without merging them into the main project.

## When collaborating, start and end clean

Before starting work on a shared github project, it's usually best to pull the most
recent version of the branch from the remote repository so you know you're working 
with the most updated code from others. Otherwise, you may find your code works 
with older version but not with the most recent versions.

Likewise, when you finish a work session, you should always commit changes locally,
and if you have a personal or shared `dev` branch, consider pushing the changes as
well. Because you're not developing in the main branch, you can push changes
without impacting anyone outside those actively working on your branch.

## Use pull requests to merge clean updates into main projects

When an update is clean and debugged, and you think it's ready to finalize, you can create a **pull request** on the `github` web interface and ask to merge changes
from your branch into the main branch of the repository. 

A large repository may have dozens of working development branches, with different
people working on each. For collaborative projects, it is usually the responsibility
of a small number of maintainers or coordinators to manage pull requests by
verifying they won't have negative impact and either approving, rejecting, or asking
for changes.
